refactor: move snippetz types to @scalar/types
