+++ packages/api-reference/src/types.ts
@@ -1,7 +1,7 @@
-import type { HarRequest } from '@scalar/snippetz/types'
 import type { ThemeId } from '@scalar/themes'
 import type { ApiReferenceConfiguration } from '@scalar/types/api-reference'
 import type { ContentType, ReferenceConfiguration, Spec } from '@scalar/types/legacy'
+import type { HarRequest } from '@scalar/types/snippetz'
 import type { Slot } from 'vue'
 
 export type { ApiReferenceConfiguration }
