+++ packages/snippetz/src/plugins/c/libcurl/libcurl.ts
@@ -1,6 +1,6 @@
 import { libcurl } from '@/httpsnippet-lite/esm/targets/c/libcurl/client'
-import type { Plugin } from '@/types'
 import { convertWithHttpSnippetLite } from '@/utils/convertWithHttpSnippetLite'
+import type { Plugin } from '@scalar/types/snippetz'
 
 /**
  * c/libcurl
