+++ packages/snippetz/src/plugins/clojure/clj_http/clj_http.ts
@@ -1,6 +1,6 @@
 import { clj_http } from '@/httpsnippet-lite/esm/targets/clojure/clj_http/client'
-import type { Plugin } from '@/types'
 import { convertWithHttpSnippetLite } from '@/utils/convertWithHttpSnippetLite'
+import type { Plugin } from '@scalar/types/snippetz'
 
 /**
  * clojure/clj_http
