+++ packages/snippetz/src/plugins/csharp/restsharp/restsharp.ts
@@ -1,6 +1,6 @@
 import { restsharp } from '@/httpsnippet-lite/esm/targets/csharp/restsharp/client'
-import type { Plugin } from '@/types'
 import { convertWithHttpSnippetLite } from '@/utils/convertWithHttpSnippetLite'
+import type { Plugin } from '@scalar/types/snippetz'
 
 /**
  * csharp/restsharp
