+++ packages/snippetz/src/plugins/java/nethttp/nethttp.ts
@@ -1,6 +1,6 @@
 import { nethttp } from '@/httpsnippet-lite/esm/targets/java/nethttp/client'
-import type { Plugin } from '@/types'
 import { convertWithHttpSnippetLite } from '@/utils/convertWithHttpSnippetLite'
+import type { Plugin } from '@scalar/types/snippetz'
 
 /**
  * java/nethttp
