+++ packages/snippetz/src/plugins/java/unirest/unirest.ts
@@ -1,6 +1,6 @@
 import { unirest } from '@/httpsnippet-lite/esm/targets/java/unirest/client'
-import type { Plugin } from '@/types'
 import { convertWithHttpSnippetLite } from '@/utils/convertWithHttpSnippetLite'
+import type { Plugin } from '@scalar/types/snippetz'
 
 /**
  * java/unirest
