+++ packages/snippetz/src/plugins/js/axios/axios.ts
@@ -1,6 +1,6 @@
 import { axios } from '@/httpsnippet-lite/esm/targets/javascript/axios/client'
-import type { Plugin } from '@/types'
 import { convertWithHttpSnippetLite } from '@/utils/convertWithHttpSnippetLite'
+import type { Plugin } from '@scalar/types/snippetz'
 
 /**
  * js/axios
