+++ packages/snippetz/src/plugins/js/xhr/xhr.ts
@@ -1,6 +1,6 @@
 import { xhr } from '@/httpsnippet-lite/esm/targets/javascript/xhr/client'
-import type { Plugin } from '@/types'
 import { convertWithHttpSnippetLite } from '@/utils/convertWithHttpSnippetLite'
+import type { Plugin } from '@scalar/types/snippetz'
 
 /**
  * js/xhr
