+++ packages/snippetz/src/plugins/kotlin/okhttp/okhttp.ts
@@ -1,6 +1,6 @@
 import { okhttp } from '@/httpsnippet-lite/esm/targets/kotlin/okhttp/client'
-import type { Plugin } from '@/types'
 import { convertWithHttpSnippetLite } from '@/utils/convertWithHttpSnippetLite'
+import type { Plugin } from '@scalar/types/snippetz'
 
 /**
  * kotlin/okhttp
