+++ packages/snippetz/src/plugins/node/axios/axios.ts
@@ -1,6 +1,6 @@
 import { axios } from '@/httpsnippet-lite/esm/targets/node/axios/client'
-import type { Plugin } from '@/types'
 import { convertWithHttpSnippetLite } from '@/utils/convertWithHttpSnippetLite'
+import type { Plugin } from '@scalar/types/snippetz'
 
 /**
  * node/axios
