+++ packages/snippetz/src/plugins/objc/nsurlsession/nsurlsession.ts
@@ -1,6 +1,6 @@
 import { nsurlsession } from '@/httpsnippet-lite/esm/targets/objc/nsurlsession/client'
-import type { Plugin } from '@/types'
 import { convertWithHttpSnippetLite } from '@/utils/convertWithHttpSnippetLite'
+import type { Plugin } from '@scalar/types/snippetz'
 
 /**
  * objc/nsurlsession
