+++ packages/snippetz/src/plugins/ocaml/cohttp/cohttp.ts
@@ -1,6 +1,6 @@
 import { cohttp } from '@/httpsnippet-lite/esm/targets/ocaml/cohttp/client'
-import type { Plugin } from '@/types'
 import { convertWithHttpSnippetLite } from '@/utils/convertWithHttpSnippetLite'
+import type { Plugin } from '@scalar/types/snippetz'
 
 /**
  * ocaml/cohttp
