+++ packages/snippetz/src/plugins/php/curl/curl.ts
@@ -1,6 +1,6 @@
 import { curl } from '@/httpsnippet-lite/esm/targets/php/curl/client'
-import type { Plugin } from '@/types'
 import { convertWithHttpSnippetLite } from '@/utils/convertWithHttpSnippetLite'
+import type { Plugin } from '@scalar/types/snippetz'
 
 /**
  * php/curl
