+++ packages/snippetz/src/plugins/php/guzzle/guzzle.ts
@@ -1,6 +1,6 @@
 import { guzzle } from '@/httpsnippet-lite/esm/targets/php/guzzle/client'
-import type { Plugin } from '@/types'
 import { convertWithHttpSnippetLite } from '@/utils/convertWithHttpSnippetLite'
+import type { Plugin } from '@scalar/types/snippetz'
 
 /**
  * php/guzzle
