+++ packages/snippetz/src/plugins/powershell/webrequest/webrequest.ts
@@ -1,6 +1,6 @@
 import { webrequest } from '@/httpsnippet-lite/esm/targets/powershell/webrequest/client'
-import type { Plugin } from '@/types'
 import { convertWithHttpSnippetLite } from '@/utils/convertWithHttpSnippetLite'
+import type { Plugin } from '@scalar/types/snippetz'
 
 /**
  * powershell/webrequest
