+++ packages/snippetz/src/plugins/shell/wget/wget.ts
@@ -1,6 +1,6 @@
 import { wget } from '@/httpsnippet-lite/esm/targets/shell/wget/client'
-import type { Plugin } from '@/types'
 import { convertWithHttpSnippetLite } from '@/utils/convertWithHttpSnippetLite'
+import type { Plugin } from '@scalar/types/snippetz'
 
 /**
  * shell/wget
