+++ packages/snippetz/src/utils/convertWithHttpSnippetLite.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from 'vitest'
 
-import type { HarRequest } from '@/types'
+import type { HarRequest } from '@scalar/types/snippetz'
 import { convertWithHttpSnippetLite } from './convertWithHttpSnippetLite'
 
 describe('convertWithHttpSnippetLite', () => {
