--- packages/types/src/external/snippetz.ts
@@ -1,1 +1,0 @@
-export type { HarRequest, TargetId, ClientId } from '@scalar/snippetz/types'
